ENH slight refactoring of `ParameterSet` hierarchy + addition of `ExplicitParameterSet`, which is FIFO
